+++ RapidFTR-Android/res/layout/child_list.xml
@@ -2,6 +2,6 @@
 
 <ListView
         android:id="@+id/child_list"
-        android:layout_height="wrap_content" android:layout_width="fill_parent"
+        android:layout_height="match_parent" android:layout_width="fill_parent"
         xmlns:android="http://schemas.android.com/apk/res/android">
 </ListView>
