+++ RapidFTR-Android/src/main/java/com/rapidftr/activity/ViewEnquiryActivity.java
@@ -18,7 +18,6 @@ import com.rapidftr.service.EnquirySyncService;
 import com.rapidftr.service.LogOutService;
 import com.rapidftr.task.AsyncTaskWithDialog;
 import com.rapidftr.task.SyncSingleRecordTask;
-import com.rapidftr.view.FormSectionView;
 import com.rapidftr.view.PotentialMatchesFormSectionView;
 import lombok.Cleanup;
 import org.json.JSONException;
@@ -35,10 +34,10 @@ public class ViewEnquiryActivity extends BaseEnquiryActivity {
 
     @Override
     protected void initializePager() {
-        FormSectionView potentialMatchesView = new PotentialMatchesFormSectionView(this) {
+        PotentialMatchesFormSectionView potentialMatchesView = new PotentialMatchesFormSectionView(this) {
             @Override
             protected HighlightedFieldsViewAdapter getHighlightedFieldsViewAdapter(List<BaseModel> models) {
-                return new HighlightedFieldsViewAdapter(getContext(), models, Child.CHILD_FORM_NAME, ViewChildActivity.class);
+                return new HighlightedFieldsViewAdapter(ViewEnquiryActivity.this, models, Child.CHILD_FORM_NAME, ViewChildActivity.class);
             }
         };
         getPager().setAdapter(new PotentialMatchesFormSectionPagerAdapter(formSections, getModel(), getEditable(), potentialMatchesView));
