+++ RapidFTR-Android/src/main/java/com/rapidftr/adapter/PotentialMatchesFormSectionPagerAdapter.java
@@ -5,16 +5,17 @@ import android.view.ViewGroup;
 import com.rapidftr.forms.FormSection;
 import com.rapidftr.forms.PotentialMatchesFormSection;
 import com.rapidftr.model.BaseModel;
-import com.rapidftr.view.FormSectionView;
+import com.rapidftr.view.DefaultFormSectionView;
 import com.rapidftr.view.PotentialMatchesFormSectionView;
 
 import java.util.List;
 
 public class PotentialMatchesFormSectionPagerAdapter extends FormSectionPagerAdapter {
 
-    private FormSectionView potentialMatchesView;
+    private PotentialMatchesFormSectionView potentialMatchesView;
 
-    public PotentialMatchesFormSectionPagerAdapter(List<FormSection> formSections, BaseModel baseModel, boolean editable, FormSectionView potentialMatchesView) {
+    public PotentialMatchesFormSectionPagerAdapter(List<FormSection> formSections, BaseModel baseModel, boolean editable,
+                                                   PotentialMatchesFormSectionView potentialMatchesView) {
         super(formSections, baseModel, editable);
         this.potentialMatchesView = potentialMatchesView;
     }
@@ -32,15 +33,17 @@ public class PotentialMatchesFormSectionPagerAdapter extends FormSectionPagerAda
     @Override
     public Object instantiateItem(ViewGroup container, int position) {
         FormSection formSection = formSections.get(position);
-        FormSectionView view;
+        View view = null;
 
         if (formSection instanceof PotentialMatchesFormSection) {
+            potentialMatchesView.initialize(formSection, baseModel);
             view = potentialMatchesView;
         } else {
-            view = createFormSectionView(container);
+            DefaultFormSectionView defaultFormSectionView = createFormSectionView(container);
+            defaultFormSectionView.initialize(formSection, baseModel);
+            view = defaultFormSectionView;
         }
 
-        view.initialize(formSection, baseModel);
         view.setEnabled(editable);
         container.addView(view, 0);
         return view;
